requirements.txt
# Gmail API
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0

# Database
SQLAlchemy==2.0.19
alembic==1.11.1

# Date handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.6.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Type checking
mypy==1.4.1

# Code formatting
black==23.7.0
isort==5.12.0

# Logging
structlog==23.1.0